+++ src/v1/BasicTypes/HashMaps/HashMap.ts
@@ -31,7 +31,7 @@
 // ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 //
-import { findAttributeNames } from "../Objects";
+import { findPropertyNames } from "../Objects";
 import { STOP_AT_NEXT_PROTOTYPE } from "../Prototypes";
 
 /**
@@ -62,12 +62,11 @@ export class HashMap<T> {
         target: HashMap<T>,
         callbackfn: (value: T, name: string, obj: HashMap<T>) => void
     ): void {
-        findAttributeNames(
-            target,
-            { nextPrototype: STOP_AT_NEXT_PROTOTYPE }
-        ).forEach((name: string) => {
-            callbackfn(target[name], name, target);
-        });
+        HashMap.keys(target)
+            .forEach((name: string) => {
+                callbackfn(target[name], name, target);
+            }
+        );
     }
 
     /**
@@ -121,12 +120,11 @@ export class HashMap<T> {
         target: HashMap<T>,
         callbackfn: (value: T, name: string, obj: HashMap<T>) => boolean
     ): boolean {
-        return findAttributeNames(
-            target,
-            { nextPrototype: STOP_AT_NEXT_PROTOTYPE }
-        ).some((name: string) => {
-            return callbackfn(target[name], name, target);
-        });
+        return HashMap.keys(target)
+            .some((name: string) => {
+                return callbackfn(target[name], name, target);
+            }
+        );
     }
 
     /**
@@ -147,11 +145,102 @@ export class HashMap<T> {
         target: HashMap<T>,
         callbackfn: (value: T, name: string, obj: HashMap<T>) => boolean
     ): boolean {
-        return findAttributeNames(
+        return HashMap.keys(target)
+            .every((name: string) => {
+                return callbackfn(target[name], name, target);
+            }
+        );
+    }
+
+    /**
+     * `first()` searches the given HashMap for a property that satisfies
+     * the test in the given callback function.
+     *
+     * On success, it returns a HashMap containing only the first property
+     * that satisfies the given callback function.
+     *
+     * On failure, it returns an empty HashMap.
+     *
+     * The order that we search the HashMap cannot be guaranteed.
+     *
+     * @param target
+     * the HashMap to search
+     * @param callbackfn
+     * the function to call when we iterate over `target`
+     * @returns
+     * - a HashMap containing the first matching property on success,
+     * - an empty HashMap on failure
+     */
+    public static first<T, R=T>(
+        target: HashMap<T>,
+        callbackfn: (value: T, name: string, obj: HashMap<T>) => boolean
+    ): HashMap<R> {
+        // our return value
+        const retval: HashMap<R> = {};
+
+        // do we have a property to return?
+        const maybePropName = HashMap.firstKey(target, callbackfn);
+        if (maybePropName) {
+            // yes we do
+            retval[maybePropName] = (target[maybePropName] as unknown) as R;
+        }
+
+        // all done
+        return retval;
+    }
+
+    /**
+     * `firstKey()` searches the given HashMap for a property that
+     * satisfies the test in the given callback function.
+     *
+     * On success, it returns the name of the first property that satisfies
+     * the given callback function.
+     *
+     * On failure, it returns null.
+     *
+     * The order that we search the HashMap cannot be guaranteed.
+     *
+     * @param target
+     * the HashMap to search
+     * @param callbackfn
+     * the function to call when we iterate over `target`
+     * @returns
+     * - a HashMap containing the first matching property on success,
+     * - an empty HashMap on failure
+     */
+    public static firstKey<T>(
+        target: HashMap<T>,
+        callbackfn: (value: T, name: string, obj: HashMap<T>) => boolean
+    ): string|null {
+        const propNames = HashMap.keys(target);
+
+        for(const name of propNames) {
+            if (callbackfn(target[name], name, target)) {
+                return name;
+            }
+        }
+
+        // if we get here, we did not find a match
+        return null;
+    }
+
+    /**
+     * `keys()` returns a list of all the property names from the given
+     * HashMap.
+     *
+     * Unlike {@link Object.keys}, `HashMap.keys()` only returns properties
+     * that belong to `target` (ie nothing that belongs to its prototypes).
+     *
+     * @param target
+     * the HashMap to example
+     * @returns
+     * the (possibly empty) list of attribute names from `target`
+     */
+    public static keys<T>(target: HashMap<T>): string[]
+    {
+        return findPropertyNames(
             target,
             { nextPrototype: STOP_AT_NEXT_PROTOTYPE }
-        ).every((name: string) => {
-            return callbackfn(target[name], name, target);
-        });
+        );
     }
 }